git-svn-id: https://phys12svn.jlab.org/repos/JANA@2402 c5ed4466-e916-0410-8347-b3263e9c103d

+++ src/JANA/JFactory.h
@@ -211,8 +211,8 @@ jerror_t JFactory<T>::Get(vector<const T*> &d)
 		throw JException(mess);
 	}	
 	// Grab the current event and run numbers
-	int event_number = eventLoop->GetJEvent().GetEventNumber();
-	int run_number = eventLoop->GetJEvent().GetRunNumber();
+	uint64_t event_number = eventLoop->GetJEvent().GetEventNumber();
+	int32_t run_number = eventLoop->GetJEvent().GetRunNumber();
 	
 	// Make sure we're initialized
 	if(!init_called){
